Refactor batch CLI into module and tighten regression semantics
